+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2022-09-26
+### Added
+- Warmup-1 icyHot
+- Warmup-1 in1020
+ 
 ## [1.1.0] - 2022-09-25
 ### Added
 - Warmup-1 missingChar